+++ src/lib/components/index.ts
@@ -1,6 +1,7 @@
 export * from './user-menu';
 export * from './notification';
 export * from './toast';
+export * from './button';
 export * from './select-list';
 export * from './paginator';
 export * from './data-table';