+++ src/lib/core/models/index.ts
@@ -2,6 +2,7 @@ export * from './notification.models';
 export * from './user-menu.models';
 export * from './settings.models';
 export * from './toast.models';
+export * from './button.models';
 export * from './select-list.models';
 export * from './data-table.models';
 export * from './stepper.models';